# nam_viet_erp_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/inventory/deduct": {
            "post": {
                "description": "Trừ tồn kho theo thuật toán FEFO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Trừ tồn kho FEFO",
                "parameters": [
                    {
                        "description": "Stock deduction request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.DeductStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/inventory/receipt": {
            "post": {
                "description": "Nhập kho sản phẩm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Nhập kho",
                "parameters": [
                    {
                        "description": "Stock receipt request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateReceiptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/inventory/validate": {
            "post": {
                "description": "Validate if stock is available for given items and warehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Kiểm tra tồn kho",
                "parameters": [
                    {
                        "description": "Stock validation request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ValidateStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateReceiptRequest": {
            "type": "object",
            "required": [
                "items",
                "warehouse_id"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/domain.ReceiptItem"
                    }
                },
                "note": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.DeductStockRequest": {
            "type": "object",
            "required": [
                "items",
                "warehouse_id"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/domain.ValidateStockItem"
                    }
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.ReceiptItem": {
            "type": "object",
            "required": [
                "batch_code",
                "expiry_date",
                "product_id",
                "quantity"
            ],
            "properties": {
                "batch_code": {
                    "type": "string"
                },
                "expiry_date": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "unit_price": {
                    "type": "number"
                }
            }
        },
...
        "domain.ValidateStockItem": {
            "type": "object",
            "required": [
                "product_id",
                "quantity",
                "uom"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "uom": {
                    "description": "The requested unit of measurement",
                    "type": "string"
                }
            }
        },
        "domain.ValidateStockRequest": {
            "type": "object",
            "required": [
                "items",
                "warehouse_id"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/domain.ValidateStockItem"
                    }
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        }
    }
}